Crate sel4_config

Modules

consts

The kernel configuration as const items.

Macros

sel4_cfg_bool

Like core::cfg!, except using the seL4 kernel configuration.

sel4_cfg_if

Like cfg_if::cfg_if!, except with sel4_cfg instead of #[cfg].

sel4_cfg_str

Like core::cfg!, except using the seL4 kernel configuration.

sel4_cfg_usize

Like core::cfg!, except using the seL4 kernel configuration.

sel4_cfg_word

Like core::cfg!, except using the seL4 kernel configuration.

sel4_cfg_wrap_match

Like sel4_cfg_match, except it works on stable, at the expense of not being an attribute macro.

Attribute Macros

sel4_cfg

Make the attached code conditional on a seL4 kernel configuration expression.

sel4_cfg_attr

Make the associated attribute expression conditional on a seL4 kernel configuration expression.

sel4_cfg_enum

Allows an enum’s variants to use the sel4_cfg attribute.

sel4_cfg_match

Allows a match expression’s variants to use the sel4_cfg attribute.

sel4_cfg_struct

Allows a struct’s fields to use the sel4_cfg attribute.